+++ morpheus-graphql-client/src/Data/Morpheus/Client.hs
@@ -1,8 +1,6 @@
 {-# LANGUAGE DataKinds #-}
 {-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE GADTs #-}
-{-# LANGUAGE NamedFieldPuns #-}
-{-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE RankNTypes #-}
 {-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE NoImplicitPrelude #-}
@@ -29,16 +27,14 @@ module Data.Morpheus.Client
     defineByIntrospectionFile,
     defineByIntrospectionFile',
     Request (..),
+    ClientResult,
     request,
     forEach,
     single,
     ResponseStream,
-    ClientResult,
   )
 where
 
-import Data.Aeson
-import qualified Data.Aeson as A
 import Data.ByteString.Lazy (ByteString)
 import qualified Data.ByteString.Lazy as L
   ( readFile,
@@ -55,19 +51,19 @@ import Data.Morpheus.Client.Declare
 import Data.Morpheus.Client.Fetch
   ( Fetch (..),
     Request (..),
-    decodeResponse,
-    processResponse,
-    toRequest,
+  )
+import Data.Morpheus.Client.Fetch.ResponseStream
+  ( ResponseStream,
+    forEach,
+    request,
+    single,
   )
 import Data.Morpheus.Client.Internal.Types
   ( ClientResult,
     ExecutableSource,
     FetchError (..),
-    HTTP,
     SchemaSource (..),
   )
-import Data.Morpheus.Client.Schema.JSON.Types
-import Data.Morpheus.Subscriptions.Internal (ApolloSubscription (..))
 import Data.Morpheus.Types.GQLScalar
   ( DecodeScalar (..),
     EncodeScalar (..),
@@ -76,27 +72,12 @@ import Data.Morpheus.Types.ID (ID (..))
 import Data.Morpheus.Types.Internal.AST
   ( ScalarValue (..),
   )
-import qualified Data.Text as T
 import Language.Haskell.TH
 import Language.Haskell.TH.Quote (QuasiQuoter)
 import Language.Haskell.TH.Syntax
   ( qAddDependentFile,
   )
-import Network.HTTP.Req
-  ( Option,
-    POST (..),
-    ReqBodyLbs (..),
-    defaultHttpConfig,
-    header,
-    lbsResponse,
-    req,
-    responseBody,
-    runReq,
-    useURI,
-  )
-import Network.WebSockets (ClientApp, receiveData, runClient, sendTextData)
 import Relude hiding (ByteString)
-import Text.URI (Authority (..), RText, RTextLabel (PathPiece), URI (..), mkURI, unRText)
 
 {-# DEPRECATED gql "use raw" #-}
 gql :: QuasiQuoter
@@ -139,87 +120,3 @@ defineByDocument doc = internalLegacyLocalDeclareTypes (GQL <$> doc)
 {-# DEPRECATED defineByIntrospection "use clientTypeDeclarations" #-}
 defineByIntrospection :: IO ByteString -> ExecutableSource -> Q [Dec]
 defineByIntrospection doc = internalLegacyLocalDeclareTypes (JSON <$> doc)
-
-headers :: Network.HTTP.Req.Option scheme
-headers = header "Content-Type" "application/json"
-
-post :: URI -> ByteString -> IO ByteString
-post uri body = case useURI uri of
-  Nothing -> fail ("Invalid Endpoint: " <> show uri <> "!")
-  (Just (Left (u, o))) -> responseBody <$> runReq defaultHttpConfig (req POST u (ReqBodyLbs body) lbsResponse (o <> headers))
-  (Just (Right (u, o))) -> responseBody <$> runReq defaultHttpConfig (req POST u (ReqBodyLbs body) lbsResponse (o <> headers))
-
-parseURI :: MonadFail m => String -> m URI
-parseURI url = maybe (fail ("Invalid Endpoint: " <> show url <> "!")) pure (mkURI (T.pack url))
-
-toPort :: Maybe Word -> Int
-toPort Nothing = 80
-toPort (Just x) = fromIntegral x
-
-getPath :: Maybe (Bool, NonEmpty (RText 'PathPiece)) -> String
-getPath (Just (_, h :| t)) = T.unpack $ T.intercalate "/" $ fmap unRText (h : t)
-getPath _ = ""
-
-useWS :: URI -> ClientApp () -> IO ()
-useWS URI {uriScheme = Just scheme, uriAuthority = Right Authority {authHost, authPort}, uriPath} app
-  | unRText scheme == "ws" = do
-    let rPath = getPath uriPath
-    let rPort = toPort authPort
-    let rHost = handleHost $ unRText authHost
-    runClient rHost rPort rPath app
-useWS uri _ = fail ("Invalid Endpoint: " <> show uri <> "!")
-
-handleHost :: Text -> String
-handleHost "localhost" = "127.0.0.1"
-handleHost x = T.unpack x
-
-requestSingle :: Fetch a => URI -> Request method a -> IO (Either (FetchError a) a)
-requestSingle uri r = decodeResponse <$> post uri (A.encode $ toRequest r)
-
-requestStream :: (Fetch a) => URI -> Request method a -> (ClientResult a -> IO ()) -> IO ()
-requestStream uri r f = useWS uri app
-  where
-    decodeMessage = (first FetchErrorParseFailure . eitherDecode) >=> processMessage
-    processMessage :: ApolloSubscription (JSONResponse a) -> Either (FetchError a) a
-    processMessage ApolloSubscription {apolloPayload = Just payload} = processResponse payload
-    processMessage ApolloSubscription {} = Left (FetchErrorParseFailure "empty message")
-    app conn = do
-      sendTextData conn (A.encode initMessage)
-      -- send initial GQL Request for subscription
-      sendTextData
-        conn
-        ( A.encode
-            ApolloSubscription
-              { apolloPayload = Just (toRequest r),
-                apolloType = "start",
-                apolloId = Just "01"
-              }
-        )
-      -- handle GQL subscription responses
-      forever $ do
-        message <- receiveData conn
-        f (decodeMessage message)
-
-initMessage :: ApolloSubscription ()
-initMessage = ApolloSubscription {apolloType = "connection_init", apolloPayload = Nothing, apolloId = Nothing}
-
--- endMessage :: Text -> ApolloSubscription ()
--- endMessage uid = ApolloSubscription {apolloType = "stop", apolloPayload = Nothing, apolloId = Just uid}
-
-data ResponseStream a = Fetch a =>
-  ResponseStream
-  { _req :: Request HTTP a,
-    _uri :: URI
-  }
-
-request :: Fetch a => String -> Args a -> IO (ResponseStream a)
-request uri requestArgs = do
-  _uri <- parseURI uri
-  let _req = Request {requestArgs}
-  pure ResponseStream {_req, _uri}
-
-forEach :: (ClientResult a -> IO ()) -> ResponseStream a -> IO ()
-forEach f ResponseStream {_uri, _req} = requestStream _uri _req f
-
-single :: ResponseStream a -> IO (ClientResult a)
-single ResponseStream {_req, _uri} = requestSingle _uri _req
